Add HTTP status code to log
For historical tracking purposes, we want to know what the HTTP status code of the server check was. In the future, we can use this attribute to check for anomalies, or revise what counts as up vs down.

However, we don't want to store JUST the HTTP status code — we also want to make a determination of up vs down, and store that in the log. I forget why as I'm committing this change, but we had a compelling reason when we talked about it last.